Update README.md
To-Do List project is an Angular application specially built to keep track of errands or tasks that need to be done. This application will be like a task keeper where the user can enter the tasks they need to do. Once they are done with their tasks they can also remove them from the list.

You can add the tasks that are to be done in a descriptive way and determine it's start and end date.
You will be able to add as many tasks as you have.
Once the task is completed, you will be able to remove it by clicking on the remove button.